pyproject.toml
[project]
name = "vibe-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.40.1",
    "finnhub-python>=2.4.23",
    "google-adk>=1.1.1",
    "jupyter>=1.1.1",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
]
